Adds unit tests for the LeftHand Exception class
The Exception class for the LeftHand client did not have any
unit tests.

This patch adds unit tests that check the output from the
from_response function and the ClientException class when an
error is passed in.

This patch also removed the use of 'ref' from the
ClientException class. It was added in fa50dedd but was
never officially released so it will not be deprecated.

Change-Id: Ifeb0687e00808155fda116ac97abcf6afaedbff5