Fixes error description and code not being saved
In LeftHand, the error exception returns messageID and message, we
are currently looking at code and desc, which do not exist. This patch
saves the correct values. This patch also throws the correct values in
the flask server as well.

There are some driver functions that look at the exact error message,
so having this saved properly is important.

Change-Id: I21399d6a7a2932485df34b2dda40c789e5edd746