Changes exception's isinstance str to basestring
The error description was not being displayed when an exception
was being raised. This was due to the patch merged in from
issue #12 (reported in the 3PAR client), which changed the body
to type bytes instead of typestr.  The check in exceptions.py is
now updated to represent this change.

Change-Id: I471192a38b47f118114239fce63c52c9764b1ac9